+++ src/logic/MacroPreferences.swift
@@ -185,6 +185,62 @@ enum ShortcutStylePreference: CaseIterable, MacroPreference {
     }
 }
 
+enum HyperKeyActionPreference: CaseIterable, MacroPreference {
+    case none
+    case leftThird
+    case rightThird
+    case leftTwoThirds
+    case rightTwoThirds
+    case restore
+
+    var localizedString: LocalizedString {
+        switch self {
+        case .none: return NSLocalizedString("Do nothing", comment: "")
+        case .leftThird: return WindowLayoutAction.leftThird.localizedTitle
+        case .rightThird: return WindowLayoutAction.rightThird.localizedTitle
+        case .leftTwoThirds: return WindowLayoutAction.leftTwoThirds.localizedTitle
+        case .rightTwoThirds: return WindowLayoutAction.rightTwoThirds.localizedTitle
+        case .restore: return WindowLayoutAction.restore.localizedTitle
+        }
+    }
+
+    var windowLayoutAction: WindowLayoutAction? {
+        switch self {
+        case .none: return nil
+        case .leftThird: return .leftThird
+        case .rightThird: return .rightThird
+        case .leftTwoThirds: return .leftTwoThirds
+        case .rightTwoThirds: return .rightTwoThirds
+        case .restore: return .restore
+        }
+    }
+}
+
+enum HyperKeyHoldDurationPreference: CaseIterable, MacroPreference {
+    case milliseconds150
+    case milliseconds200
+    case milliseconds250
+    case milliseconds300
+
+    var localizedString: LocalizedString {
+        switch self {
+        case .milliseconds150: return "150 ms"
+        case .milliseconds200: return "200 ms"
+        case .milliseconds250: return "250 ms"
+        case .milliseconds300: return "300 ms"
+        }
+    }
+
+    var seconds: TimeInterval {
+        switch self {
+        case .milliseconds150: return 0.15
+        case .milliseconds200: return 0.2
+        case .milliseconds250: return 0.25
+        case .milliseconds300: return 0.3
+        }
+    }
+}
+
 enum ShowHowPreference: CaseIterable, MacroPreference {
     case show
     case hide